+++ src/main/java/janggi/util/DatabaseInitializer.java
@@ -1,6 +1,5 @@
 package janggi.util;
 
-import janggi.repository.JdbcConnectionManager;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStream;
