+++ src/test/java/janggi/repository/DatabaseInitializerTest.java
@@ -3,6 +3,7 @@ package janggi.repository;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import janggi.util.DatabaseInitializer;
+import janggi.util.JdbcConnectionManager;
 import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
